Remove client interface
The removal is due to the fact that it is other application logic
creeping into the pkg that add no benifit. You should be handling your
specific logic for interfaces in your app. I don't see where having an
interface defined here would help. If in the future this is necessary to
make it easier to maintain I would suggest calling the interface
XsollaClient as the Client name is reserved for the actual object.
Perhaps this is backwards?